[IMPROVEMENT] board displays real words and mess.
Mess is aimed at hiding the real words but currently is just X's
the way the real words and the board is created is now different as well.

Previously, the pieces were created(new letterPiece) with the loadedWords.
Now, the pieces are created as an 8x8 grid of letterPieces.
Then the words are loaded and sections of the 8x8 grid are set to the letters
of the words. The remaining, unset letterPieces on the grid are then set to
the X's.