[IMPROVEMENT] Made easier to select letters diagonally.
Previously, trying to select diagonally would usually also select a
neighbouring letter.
Now, a buffer zone is introduced to make it easier to select letters
diagonally. This makes the intersection check area of the mouse and letter
smaller so to select a neighbouring letter as well the user has to really
hover near the actual letter and not just slightly, hover over the edge.