[REFACTOR] Encapsulate dictionary login into controller class
Previously, the dictionary logic was in its own file but in the global scope.
Now, it is in it's own "class". This exposes the public function addWords
so that the html can access it but the function validWords which is used
internally is no longer externally accessible. This is to reflect the
intention that validWords is only intended to be used inside the dictionary
controller and not by or for anything else.