[REFACTOR] Moved word loading from global scope to dictionary controller.
Previously, it was located in the game.js in a globally scoped function.
Now, it exists in the dictionary controller as it is logically related to
the dictionary. Ie, loading of the words is the dictionary controllers job
so that functionality should exist in the dictionary controller, not the
game.