[REFACTOR] Lots of changes to make it easier to have two boards
Previously, the code was made to work with one board. Lots of code related to
boards was in the game.js file.
Now, to facilate having two boards and switching between them, the logic for
boards has been refactored out of game.js into board.js. This allows for
easy re-use of the board code without having duplicate code all over the place.
It also means the board code/logic is encapsulated into the board class making
a nicer design.

The player controller and board pieces array is still coupled fairly tightly.
This will probably be next to go.