[IMPROVEMENT] Once word found on one board, switches to other board.
Previously, UI was a bit messed.
Now, starting on the left board, the user can find a word. Once a word has
been found the UI is switched to be active on the right board. Then when
a word has been found on that board, it switches back to the left. And on and
on and on and on and on.