[IMPROVEMENT] words scattered on board correctly.
Previously, the words weren't scattered on the board in any meaningful way.
They were in a simple order making them easy to find.

Now, the words are scattered on the board. They are randomly chosen to be
horizontal or vertical and then laid out along the board in one of those
orientations.